Переделать: систему хранения бонусов

+++ internal/entity/order.go
@@ -7,6 +7,6 @@ type Order struct {
 	UserID     int       `json:"user_id,omitempty"`
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    float64   `json:"accrual,omitempty"`
+	Accrual    int       `json:"accrual,omitempty"`
 	UploadedAt time.Time `json:"uploaded_at" db:"uploaded_at"`
 }
